perf(model): hold a file's size in a u64, cutting peak memory by a fifth
`File::size` was a `u128`. It is the single most repeated field in the
model — one per file, millions of them — and because it is the larger
variant it sets the size of `FileOrFolder`, which is what every slot in
every folder costs. At u128 that enum is 24 bytes; at u64 it is 16.

On a 4.23M-entry volume:

    peak RSS   824 MB -> 654 MB   (-21%)
    tree       0.73s  -> 0.68s

A u64 counts to 16 EiB. No file reaches that and no volume does, so the
range was never doing anything. `Folder::size` stays u128, where sums
across a whole tree accumulate and the width costs 379k folders rather
than 4.2M files.

Totals, entry counts, hard-link and reflink counts are unchanged.

+++ libdiskonaut/src/scan/tests.rs
@@ -370,7 +370,7 @@ fn hard_links_count_once_per_folder_when_nested() {
             .unwrap_or_else(|| panic!("{path:?} should exist"))
         {
             crate::FileOrFolder::Folder(folder) => folder.size,
-            crate::FileOrFolder::File(file) => file.size,
+            crate::FileOrFolder::File(file) => u128::from(file.size),
         }
     };
